LNK-295: Created ConceptMap and modified a supporting event to handle shifting identifier data to type within a Location
